backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
google-genai==1.0.0
psycopg2-binary==2.9.9
pydantic==2.9.2
